consolidate build system into a single package
