--- config.py
@@ -1,2 +1,0 @@
-
-devices = ['NetgateSG1100', 'UniFiUSWFlex', 'NUC10i5FNH', 'RPi4']
