+++ nimble_build_system/cad/__init__.py
@@ -1,5 +1,5 @@
 """
-On loading nimble_builder the RackParameters dataclass will be available.
+On loading nimble_build_system.cad the RackParameters dataclass will be available.
 """
 from dataclasses import dataclass
 from typing import Literal
