+++ nimble_build_system/utils/devices_json_updater.py
